[syftcli] rename src to syftcli
