+++ .isort.cfg
@@ -3,9 +3,10 @@ profile=black
 force_single_line=True
 known_syft=syft
 known_grid=grid
+known_syftcli=syftcli
 known_first_party=src
 remove_redundant_aliases=True
-sections=FUTURE,STDLIB,THIRDPARTY,SYFT,GRID,FIRSTPARTY,LOCALFOLDER
+sections=FUTURE,STDLIB,THIRDPARTY,SYFT,GRID,SYFTCLI,FIRSTPARTY,LOCALFOLDER
 lines_between_types=0
 force_sort_within_sections=True
 import_heading_future=future
@@ -13,6 +14,7 @@ import_heading_stdlib=stdlib
 import_heading_thirdparty=third party
 import_heading_syft=syft absolute
 import_heading_grid=grid absolute
+import_heading_syftcli=syftcli absolute
 import_heading_firstparty=first party
 import_heading_localfolder=relative
 ignore_comments=False
