+++ packages/syftcli/.bumpversion.cfg
@@ -5,6 +5,6 @@ tag_name = {new_version}
 commit = True
 commit_message = Bump version: {current_version} → {new_version}
 
-[bumpversion:file:src/syftcli/version.py]
+[bumpversion:file:syftcli/version.py]
 
 [bumpversion:file:setup.py]
