+++ packages/syftcli/syftcli/core/console.py
