+++ tox.ini
@@ -112,7 +112,7 @@ commands =
     ; Check if SYFT_CLI_VERSION is set or choosing the current version available
     bash -c 'if [ -z $SYFT_CLI_VERSION ]; then \
         echo "SYFT_CLI_VERSION is not set"; \
-        SYFT_CLI_VERSION=$(python3 src/syftcli/version.py); \
+        SYFT_CLI_VERSION=$(python3 syftcli/version.py); \
         echo "Setting SYFT_CLI_VERSION to $SYFT_CLI_VERSION"; \
     else \
         echo "SYFT_CLI_VERSION is already set to $SYFT_CLI_VERSION"; \
@@ -121,11 +121,11 @@ commands =
     echo "Building syftcli-$SYFT_CLI_VERSION for $OSTYPE" && \
 
     if [[ "$OSTYPE" == "darwin"* ]]; then \
-         pyinstaller --clean --onefile --name syftcli-v$SYFT_CLI_VERSION-macos-universal2 --distpath ./dist/cli src/syftcli/cli.py --target-arch universal2; \
+         pyinstaller --clean --onefile --name syftcli-v$SYFT_CLI_VERSION-macos-universal2 --distpath ./dist/cli syftcli/cli.py --target-arch universal2; \
     elif [[ "$OSTYPE" == "linux-gnu"* ]]; then \
-        pyinstaller --clean --onefile --name syftcli-v$SYFT_CLI_VERSION-linux-x86_64 --distpath ./dist/cli src/syftcli/cli.py; \
+        pyinstaller --clean --onefile --name syftcli-v$SYFT_CLI_VERSION-linux-x86_64 --distpath ./dist/cli syftcli/cli.py; \
     else \
-        pyinstaller --clean --onefile --name syftcli-v$SYFT_CLI_VERSION-windows-x86_64  --distpath ./dist/cli src/syftcli/cli.py; \
+        pyinstaller --clean --onefile --name syftcli-v$SYFT_CLI_VERSION-windows-x86_64  --distpath ./dist/cli syftcli/cli.py; \
     fi'
